adding doctor controller and correcting asynchandler

+++ Backend/src/controllers/user.controller.js
@@ -2,10 +2,15 @@ import {User} from "../models/user.model.js"
 import { asyncHandler } from "../utils/asyncHandler.js"
 import { ApiResponse } from "../utils/ApiResponse.js"
 import { ApiError } from "../utils/ApiError.js"
+import mongoose from "mongoose"
 
 
 const updateUser=asyncHandler(async(req,res)=>{
     const userId=req.params.id 
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+        throw new ApiError(400, "Invalid user ID format")
+    }
 
     const user=await User.findByIdAndUpdate(userId, {$set:req.body},{new:true}).select("-password")
 
@@ -21,10 +26,15 @@ const updateUser=asyncHandler(async(req,res)=>{
 const deleteUser=asyncHandler(async(req,res)=>{
     const userId=req.params.id
 
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+        throw new ApiError(400, "Invalid user ID format")
+    }
+
+
     const user=await User.findByIdAndDelete(userId).select("-password")
 
     if(!user){
-        throw new ApiError(400, "User not found")
+        throw new ApiError(400 , "User not found")
     }
 
     res.status(200).json(new ApiResponse(200, user,"User deleted successfully"))
@@ -34,6 +44,11 @@ const deleteUser=asyncHandler(async(req,res)=>{
 
 const getSingleUser=asyncHandler(async(req,res)=>{
     const userId=req.params.id
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+        throw new ApiError(400, "Invalid user ID format")
+    }
+
 
     const user=await User.findById(userId).select("-password")
